requirements.txt
igramscraper==0.3.5
python-arango==7.2.0
